Made code changes to improve how models are loaded in model editor
