+++ app/static/modelEditor/forms/mesh.js
@@ -1,6 +1,7 @@
 var _ = require('underscore');
 var $ = require('jquery');
 var View = require('ampersand-view');
+
 module.exports = View.extend({
     template: '<tr> \
   <td> \
