+++ app/templates/modelEditor.html
@@ -9,6 +9,12 @@
 
 <div class="modelEditor">
     <div class="span8">
+        <div class="modelLoader">
+        </div>
+        <div class="publicModelLoader">
+        </div>
+        <div class="meshLoader">
+        </div>
         <div>
             <h1>Model Editor</h1>
             <br>
